Separate componentes and containers
