+++ src/components/Button/Button.js
@@ -1,0 +1,3 @@
+import React from "react"
+
+export const Button = props => <button onClick={() => props.incrementCounter()}>Increment</button>
