+++ src/components/Header/header.less
@@ -1,0 +1,22 @@
+@textColor: #61dafb;
+
+.App-header {
+  background-color: #222;
+  height: 150px;
+  padding: 20px;
+  color: @textColor;
+}
+
+.App-logo {
+  animation: App-logo-spin infinite 20s linear;
+  height: 80px;
+}
+
+.App-title {
+  font-size: 1.5em;
+}
+
+@keyframes App-logo-spin {
+  from { transform: rotate(0deg); }
+  to { transform: rotate(360deg); }
+}
