+++ src/containers/App/App.js
@@ -1,7 +1,8 @@
 import React from "react"
-import "../style/style.less"
-import logo from "../assets/img/logo.svg"
+import { Header } from "../../components/Header/Header"
+import { Button } from "../../components/Button/Button"
 import { hot } from "react-hot-loader"
+import "./app.less"
 
 class App extends React.Component {
 	constructor() {
@@ -9,22 +10,23 @@ class App extends React.Component {
 		this.state = {
 			counter: 0
 		}
+		this.incrementCounter = this.incrementCounter.bind(this)
+	}
+
+	incrementCounter() {
+		this.setState({ counter: this.state.counter + 1 })
 	}
 
 	render() {
 		return (
 			<div className="App">
-				<header className="App-header">
-					<img src={logo} className="App-logo" alt="logo" />
-					<h1 className="App-title">Welcome to React-Less-Template</h1>
-					<h2 className="App-title">Hot Counter: {this.state.counter}</h2>
-				</header>
+				<Header counter={this.state.counter} />
 				<p className="App-intro">
 					To get finished, edit <code>src/components/app.js</code> and save to reload.
 				</p>
 				<p className="App-intro">Style, Logo and some webpack configuration taken from the Create React App team. Props to them for all their work.</p>
 				<p className="App-intro">
-					<button onClick={() => this.setState({ counter: this.state.counter + 1 })}>Increment</button>
+					<Button incrementCounter={this.incrementCounter} />
 				</p>
 			</div>
 		)
