+++ src/containers/App/app.less
@@ -1,0 +1,11 @@
+body {
+    margin: 0;
+}
+
+.App {
+    text-align: center;
+}
+
+.App-intro {
+    font-size: large;
+}
