+++ src/index.js
@@ -1,6 +1,7 @@
 import React from "react"
 import ReactDOM from "react-dom"
-import App from "./components/app"
+// import App from "./components/app"
+import App from "./containers/App/App"
 
 ReactDOM.render(<App />, document.querySelector("#root"))
