--- src/style/style.less
@@ -1,34 +1,0 @@
-@textColor: #61dafb;
-
-body {
-    margin: 0;
-}
-
-.App {
-    text-align: center;
-}
-  
-.App-logo {
-    animation: App-logo-spin infinite 20s linear;
-    height: 80px;
-}
-
-.App-header {
-    background-color: #222;
-    height: 150px;
-    padding: 20px;
-    color: @textColor;
-}
-
-.App-title {
-    font-size: 1.5em;
-}
-
-.App-intro {
-    font-size: large;
-}
-
-@keyframes App-logo-spin {
-    from { transform: rotate(0deg); }
-    to { transform: rotate(360deg); }
-}
